Compiler warning: ineffective check; promoted type so that invalid port numbers can be parsed and then checked for validity with an explicit error message, rather than "cannot read its parameters"

+++ src/Bindings/ManualBindings_Network.cpp
@@ -95,7 +95,7 @@ static int tolua_cNetwork_CreateUDPEndpoint(lua_State * L)
 	}
 
 	// Read the params:
-	UInt16 port;
+	int port;
 	cLuaState::cTableRefPtr callbacks;
 	if (!S.GetStackValues(2, port, callbacks))
 	{
